Successful file renaming
This commit includes the first successful version of this code which runs through
a given folder and renames the photo files with a prefix consisting of the date
on which the photo was taken. This will allow the user to order the files by name
which is implicitly ordering them by date taken.